# customer-service/target/classes/static/openapi.yml
openapi: 3.0.3
info:
  title: Moneybags Customer Service API
  version: 1.0.0
  description: >-
    Customer/CIF, address, KYC, risk, communication preference, nominee,
    beneficiary, eligibility, expiry-alert and transactional-outbox APIs.
    All examples are fictional demo-profile data.
servers:
  - url: http://localhost:8090
    description: API Gateway
  - url: http://localhost:8082
    description: Customer Service directly
...
paths:
  /api/v1/customers:
    get:
      tags: [Customers]
      operationId: listCustomers
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Customer'}
    post:
      tags: [Customers]
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CustomerCreateRequest'}
            example:
              firstName: Meera
              lastName: Joshi
              dob: '1994-04-12'
              gender: FEMALE
              mobile: '9898989898'
              email: meera.joshi@example.test
              panNo: PQRST6789L
              status: ACTIVE
              kycStatus: PENDING
      responses:
        '201':
          description: Customer registered with a generated CIF
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Customer'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/v1/customers/{cif}/addresses:
    get:
      tags: [Addresses]
      operationId: listAddresses
      parameters: [{$ref: '#/components/parameters/Cif'}]
      responses:
        '200':
          description: Customer addresses
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Address'}
    post:
      tags: [Addresses]
      operationId: addAddress
      parameters: [{$ref: '#/components/parameters/Cif'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddressRequest'}
            example:
              addressType: OFFICE
              line1: 15 Demo Business Park
              city: Pune
              state: Maharashtra
              pincode: '411001'
              country: India
              isCurrent: true
      responses:
        '201': {description: Address added}
...
  /api/v1/customers/{cif}/kyc-documents:
    get:
      tags: [KYC]
      operationId: listKycDocuments
      parameters: [{$ref: '#/components/parameters/Cif'}]
      responses:
        '200':
          description: KYC metadata; raw document number is never returned
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/KycDocument'}
    post:
      tags: [KYC]
      operationId: submitKycDocument
      parameters: [{$ref: '#/components/parameters/Cif'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KycSubmitRequest'}
            example:
              docType: PAN_CARD
              docNumber: XYZPS5678G
              expiryDate: '2030-07-25'
              filePath: demo://kyc/CIF900102/pan-card
      responses:
        '201': {description: KYC reference stored and number hashed}
  /api/v1/customers/{cif}/kyc-documents/{documentId}/assignment/{employeeId}:
    put:
      tags: [KYC]
      operationId: assignKycDocument
      parameters:
        - $ref: '#/components/parameters/Cif'
        - $ref: '#/components/parameters/DocumentId'
        - in: path
          name: employeeId
          required: true
          schema: {type: integer, format: int64}
          example: 1002
      responses:
        '200': {description: KYC document assigned}
  /api/v1/customers/{cif}/kyc-documents/{documentId}/decision:
    patch:
      tags: [KYC]
      operationId: decideKycDocument
      parameters:
        - $ref: '#/components/parameters/Cif'
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/KycDecisionRequest'}
            examples:
              verify:
                value: {status: VERIFIED, employeeId: 1002}
              reject:
                value: {status: REJECTED, employeeId: 1002, rejectionReason: Image is unreadable}
      responses:
        '200': {description: KYC decision saved and event added to the outbox}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/v1/customers/kyc/expiry-alerts/process:
    post:
      tags: [KYC]
      operationId: processKycExpiryAlerts
      parameters:
        - in: query
          name: asOfDate
          required: false
          schema: {type: string, format: date}
          example: '2029-03-01'
      responses:
        '200':
          description: Number of new expiry events recorded
          content:
            application/json:
              example: {alertsCreated: 1}
...
components:
  parameters:
    Cif:
      in: path
      name: cif
      required: true
      schema: {type: string}
      example: CIF900101
...
    DocumentId:
      in: path
      name: documentId
      required: true
      schema: {type: integer, format: int64}
      example: 3100
...
  responses:
...
    Conflict:
      description: Business-rule conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    CustomerStatus:
      type: string
      enum: [ACTIVE, INACTIVE, DECEASED, BLOCKED]
    CustomerCreateRequest:
      type: object
      required: [firstName, dob, gender, mobile, panNo, status, kycStatus]
      properties:
        userId: {type: integer, format: int64, nullable: true}
        firstName: {type: string, maxLength: 80}
        lastName: {type: string, maxLength: 80}
        dob: {type: string, format: date}
        gender: {type: string, enum: [MALE, FEMALE, NON_BINARY, UNDISCLOSED]}
        mobile: {type: string, pattern: '^[6-9][0-9]{9}$'}
        email: {type: string, format: email}
        panNo: {type: string, pattern: '^[A-Z]{5}[0-9]{4}[A-Z]$'}
        status: {$ref: '#/components/schemas/CustomerStatus'}
        kycStatus: {type: string, enum: [PENDING], description: Registration always starts as PENDING.}
...
    Customer:
      type: object
      properties:
        cifNo: {type: string, example: CIF900101}
        userId: {type: integer, format: int64, nullable: true}
        relationshipManagerEmpId: {type: integer, format: int64, nullable: true}
        firstName: {type: string}
        lastName: {type: string}
        dob: {type: string, format: date}
        gender: {type: string, enum: [MALE, FEMALE, NON_BINARY, UNDISCLOSED]}
        mobile: {type: string}
        email: {type: string, format: email}
        panNo: {type: string}
        status: {$ref: '#/components/schemas/CustomerStatus'}
        kycStatus: {type: string, enum: [PENDING, VERIFIED, REJECTED, EXPIRED]}
        riskClassification: {type: string, enum: [LOW, MEDIUM, HIGH]}
        preferredCommunicationChannel: {type: string, enum: [EMAIL, SMS, PUSH, NONE]}
        emailNotificationsEnabled: {type: boolean}
        smsNotificationsEnabled: {type: boolean}
        pushNotificationsEnabled: {type: boolean}
        kycFailureCount: {type: integer}
    AddressRequest:
      type: object
      required: [addressType, line1, city, state, pincode, country, isCurrent]
      properties:
        addressType: {type: string, enum: [RESIDENTIAL, PERMANENT, OFFICE]}
        line1: {type: string}
        city: {type: string}
        state: {type: string}
        pincode: {type: string}
        country: {type: string}
        isCurrent: {type: boolean}
    Address:
      allOf:
        - $ref: '#/components/schemas/AddressRequest'
        - type: object
          properties:
            addressId: {type: integer, format: int64}
    KycSubmitRequest:
      type: object
      required: [docType, docNumber, filePath]
      properties:
        docType: {type: string}
        docNumber: {type: string, writeOnly: true}
        expiryDate: {type: string, format: date, nullable: true}
        filePath: {type: string, description: Opaque secure-storage reference}
    KycDecisionRequest:
      type: object
      required: [status, employeeId]
      properties:
        status: {type: string, enum: [VERIFIED, REJECTED]}
        employeeId: {type: integer, format: int64}
        rejectionReason: {type: string, nullable: true}
    KycDocument:
      type: object
      properties:
        docId: {type: integer, format: int64}
        docType: {type: string}
        documentNumberHash: {type: string, minLength: 64, maxLength: 64}
        expiryDate: {type: string, format: date, nullable: true}
        filePath: {type: string}
        verifyStatus: {type: string, enum: [PENDING, VERIFIED, REJECTED]}
        assignedToEmpId: {type: integer, format: int64, nullable: true}
        verifiedByEmpId: {type: integer, format: int64, nullable: true}
        rejectionReason: {type: string, nullable: true}
        submittedAt: {type: string, format: date-time}
        verifiedAt: {type: string, format: date-time, nullable: true}
        expiryAlertedAt: {type: string, format: date-time, nullable: true}
...
    Error:
      type: object
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        error: {type: string}
        message: {type: string}
        path: {type: string}